+++ changes.md
@@ -1,3 +1,7 @@
+### 1.0.3
+* Added the ability to set commentChar when loading template from file.
+  Any line that starts with the commentChar will be ignore when building the commit message
+
 ### 1.0.2
 * Added the ability to load template from file